# docs/openapi.yaml
...
openapi: 3.0.0
info:
  title: Persons Endpoint
  description: Persons API for the udaconnect project
  version: 1.0.0
paths:
...
  /api/persons/{person_id}/connection:
    get:
      summary: Gets the connections of a specific Person
      description: |
        Retrieves all connections for a specific Person
      operationId: getConnections
      parameters:
      - in: query
        description: Proximity to a given user in meters
        name: distance
        required: True
        schema:
          type: string
      - in: query
        description: Start date range
        name: startDate
        required: True
        schema:
          type: string
      - in: query
        description: End date range
        name: endDate
        required: True
        schema:
          type: string
      - name: person_id
        in: path
        description: The Person's ID
        required: true
        schema:
          type: string
      responses:
        "200":
          description: Successfully retrieved all connections for a Person
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Connection'
        "400":
          description: Bad input parameter
...
components:
  schemas:
...
    Connection:
      required:
      - person
      - location
      type: object
      properties:
        person:
          type: object
          example:
          - id: 7
            first_name: "Harry"
            last_name: "Truman"
            company_name: "Target"
        location:
          type: object
          example:
          - id: 4
            person_id: 4
            longitude: 44.733443
            latitude: -321.46224
            creation_time: "2022-03-21 11:20:05"
...